[SYCL][Vitis] Remove terminate_xsimk from the SYCL runtime.
add the sources for a utility that fixes xsim issues if needed by the users.
