+++ sycl/plugins/opencl/pi_opencl.cpp
@@ -18,7 +18,6 @@
 
 #include <CL/sycl/detail/cl.h>
 #include <CL/sycl/detail/pi.h>
-#include <CL/sycl/detail/terminate_xsimk.hpp>
 
 #include <algorithm>
 #include <cassert>
@@ -1435,8 +1434,6 @@ pi_result piextKernelGetNativeHandle(pi_kernel kernel,
 pi_result piTearDown(void *PluginParameter) {
   (void)PluginParameter;
 
-  /// cleanup the potential left-overs from the hw_emu simulator
-  terminate_xsimk();
   return PI_SUCCESS;
 }
